Rename make-profile.js to processed-profile.js
