+++ src/test/components/LocalTrack.test.js
@@ -24,7 +24,7 @@ import {
 } from '../../selectors/profile';
 import { getSelectedThreadIndex } from '../../selectors/url-state';
 import mockCanvasContext from '../fixtures/mocks/canvas-context';
-import { getNetworkTrackProfile } from '../fixtures/profiles/make-profile';
+import { getNetworkTrackProfile } from '../fixtures/profiles/processed-profile';
 import { getProfileWithNiceTracks } from '../fixtures/profiles/tracks';
 import { storeWithProfile } from '../fixtures/stores';
