+++ src/test/components/TrackScreenshots.test.js
@@ -26,7 +26,7 @@ import {
   addRootOverlayElement,
   removeRootOverlayElement,
 } from '../fixtures/utils';
-import { getScreenshotTrackProfile } from '../fixtures/profiles/make-profile';
+import { getScreenshotTrackProfile } from '../fixtures/profiles/processed-profile';
 import { getProfileWithNiceTracks } from '../fixtures/profiles/tracks';
 
 // Mock out the getBoundingBox to have a 400 pixel width.
