+++ src/test/fixtures/stores.js
@@ -7,7 +7,7 @@ import createStore from '../../app-logic/create-store';
 import { viewProfile } from '../../actions/receive-profile';
 import { createGeckoProfileWithJsTimings } from './profiles/gecko-profile';
 import { processProfile } from '../../profile-logic/process-profile';
-import { getProfileFromTextSamples } from './profiles/make-profile';
+import { getProfileFromTextSamples } from './profiles/processed-profile';
 
 import type { Store } from '../../types/store';
 import type { Profile } from '../../types/profile';
