+++ src/test/store/actions.test.js
@@ -15,7 +15,7 @@ import {
   changeSelectedCallNode,
   changeShowJsTracerSummary,
 } from '../../actions/profile-view';
-import { getProfileFromTextSamples } from '../fixtures/profiles/make-profile';
+import { getProfileFromTextSamples } from '../fixtures/profiles/processed-profile';
 import { selectedThreadSelectors } from '../../selectors/per-thread';
 
 describe('selectors/getStackTimingByDepth', function() {
